git-svn-id: svn+ssh://svn.cern.ch/reps/atlasusr/end/rootpy/trunk@25541 4525493e-7705-40b1-a816-d608a930855b

+++ rootpy/classfactory.py
@@ -1,9 +1,14 @@
 import ROOT
 import uuid
 import os
+import atexit
+
+__classes = {}
 
 def make_class(declaration, headers=None):
-
+    
+    if __classes.has_key(declaration):
+        return True
     source = ""
     if headers:
         if type(headers) is not list:
@@ -11,19 +16,31 @@ def make_class(declaration, headers=None):
         for header in headers:
             source += "#include %s\n"% header
     source += "#ifdef __CINT__\n"
-    source += "#pragma link C++ class %s;\n"% declaration
+    source += "#pragma link C++ class %s+;\n"% declaration
     source += "#else\n"
     source += "using namespace std;\n"
     source += "template class %s;\n"% declaration
     source += "#endif\n"
-    tmpfilename = "%s.h"% uuid.uuid4().hex
+    tmpfilename = "%s.C"% uuid.uuid4().hex
     tmpfile = open(tmpfilename,'w')
     tmpfile.write(source)
-    level = ROOT.gErrorIgnoreLevel
-    #ROOT.gErrorIgnoreLevel = ROOT.kFatal
-    ROOT.gROOT.ProcessLine(".L %s+"% tmpfilename)
-    print ROOT.vector("vector<int>")
-    #os.unlink(tmpfilename)
-    os.unlink("%s.d"% tmpfilename.replace('.','_'))
-    os.unlink("%s.so"% tmpfilename.replace('.','_'))
-    ROOT.gErrorIgnoreLevel = level
+    tmpfile.close()
+    msg_level = ROOT.gErrorIgnoreLevel
+    ROOT.gErrorIgnoreLevel = ROOT.kFatal
+    success = ROOT.gROOT.ProcessLine(".L %s+"% tmpfilename) == 0
+    ROOT.gErrorIgnoreLevel = msg_level
+    if success:
+        __classes[declaration] = tmpfilename
+    else:
+        os.unlink(tmpfilename)
+        os.unlink("%s.d"% tmpfilename.replace('.','_'))
+        os.unlink("%s.so"% tmpfilename.replace('.','_'))
+    return success
+
+@atexit.register
+def __cleanup():
+    
+    for tmpfilename in __classes.values():
+        os.unlink(tmpfilename)
+        os.unlink("%s.d"% tmpfilename.replace('.','_'))
+        os.unlink("%s.so"% tmpfilename.replace('.','_'))
